Replace int with ull to avoid overflow

+++ Mathematics/NumberTheory/trial_division_primality_test.cpp
@@ -25,7 +25,7 @@ bool prime(ull n){
     if(n % 2 == 0){
         return false;
     }
-    for(int d = 3; d * d <= n; d += 2){
+    for(ull d = 3; d * d <= n; d += 2){
         if(n % d == 0){
             return false;
         }
